Dynamic IP Address Service Server Errors
Error model
Processing the errors on the server side

+++ 7. Dynamic_IP_Address_Service_Server/Dynamic_IP_Address_Service_Server.BLL/Controllers/DomainController.cs
@@ -1,40 +1,36 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
-using System.Net.Http;
 using System.Web.Http;
 using AutoMapper;
 using Dynamic_IP_Address_Service_Server.BLL.DTO;
+using Dynamic_IP_Address_Service_Server.BLL.Filters;
 using Dynamic_IP_Address_Service_Server.DAL.Context;
 using Dynamic_IP_Address_Service_Server.DAL.Infrastructure;
 using Dynamic_IP_Address_Service_Server.DAL.Models;
+using Dynamic_IP_Address_Service_Server.DAL.Models.Error;
 using Newtonsoft.Json;
 
 namespace Dynamic_IP_Address_Service_Server.BLL.Controllers
 {
+    [Exception]
     public class DomainController : ApiController
     {
         [HttpPost]
         public IHttpActionResult Add(DomainDTO domainDto)
         {
             var domain = Mapper.Map<Domain>(domainDto);
-            Domain result;
             using (IUnitOfWork uow = new UnitOfWork(new EntityContext()))
             {
                 if (uow.DomainRepository.CheckForUniq(domain))
                 {
+                    if (uow.UserRepository.GetByLogin(domainDto.Login) == null)
+                        throw ErrorConsts.USER_NOT_FOUND;
                     domain.User = uow.UserRepository.GetByLogin(domainDto.Login);
                     uow.DomainRepository.Insert(domain);
-                    uow.Commit();
-                    result = uow.DomainRepository.GetByDomainName(domainDto.Domain); // the domain haven't get common filds before insert to table
+                    uow.Commit(); // the domain haven't get common filds before insert to table
+                    return Ok(JsonConvert.SerializeObject(uow.DomainRepository.GetByDomainName(domainDto.Domain)));
                 }
-                else
-                {
-                    return Conflict();
-                }
+                throw ErrorConsts.DOMAIN_UNAVAILIBLE;
             }
-            return Ok(JsonConvert.SerializeObject(result));
         }
 
         [HttpPost]
@@ -46,17 +42,16 @@ namespace Dynamic_IP_Address_Service_Server.BLL.Controllers
                 if (uow.DomainRepository.CheckForUniq(domain))
                 {
                     var original = uow.DomainRepository.GetById(Guid.Parse(editDomainDto.Id));
+                    if (original == null)
+                        throw ErrorConsts.DOMAIN_NOT_FOUND;
                     original.DomainName = editDomainDto.Domain;
                     original.Ip = editDomainDto.Ip;
                     uow.DomainRepository.Update(original);
                     uow.Commit();
+                    return Ok();
                 }
-                else
-                {
-                    return Conflict();
-                }
+                throw ErrorConsts.DOMAIN_UNAVAILIBLE;
             }
-            return Ok();
         }
 
         [HttpDelete]
@@ -64,6 +59,8 @@ namespace Dynamic_IP_Address_Service_Server.BLL.Controllers
         {
             using (IUnitOfWork uow = new UnitOfWork(new EntityContext()))
             {
+                if (uow.DomainRepository.GetById(Guid.Parse(id)) == null)
+                    throw ErrorConsts.DOMAIN_NOT_FOUND;
                 uow.DomainRepository.DeleteById(Guid.Parse(id));
                 uow.Commit();
             }
